refactor: move custom rpc logic to a separate module
