+++ state-chain/node/src/service.rs
@@ -26,7 +26,7 @@ use sc_service::{error::Error as ServiceError, Configuration, TaskManager, WarpS
 use sc_telemetry::{Telemetry, TelemetryWorker};
 use sc_transaction_pool_api::OffchainTransactionPoolFactory;
 use sp_consensus_aura::sr25519::AuthorityPair as AuraPair;
-use std::{marker::PhantomData, sync::Arc, time::Duration};
+use std::{sync::Arc, time::Duration};
 
 use custom_rpc::{monitoring::MonitoringApiServer, CustomApiServer, CustomRpc};
 use state_chain_runtime::{self, opaque::Block, RuntimeApi};
@@ -303,20 +303,18 @@ pub fn new_full<
 				)?;
 
 				// Implement custom RPC extensions
-				module.merge(CustomApiServer::into_rpc(CustomRpc {
-					client: client.clone(),
-					backend: backend.clone(),
-					_phantom: PhantomData,
-					executor: executor.clone(),
-				}))?;
+				module.merge(CustomApiServer::into_rpc(CustomRpc::new(
+					client.clone(),
+					backend.clone(),
+					executor.clone(),
+				)))?;
 
 				// Implement custom RPC extensions
-				module.merge(MonitoringApiServer::into_rpc(CustomRpc {
-					client: client.clone(),
-					backend: backend.clone(),
-					_phantom: PhantomData,
-					executor: executor.clone(),
-				}))?;
+				module.merge(MonitoringApiServer::into_rpc(CustomRpc::new(
+					client.clone(),
+					backend.clone(),
+					executor.clone(),
+				)))?;
 
 				Ok(module)
 			};
